# Generalized LU factorization