This PR implements support for generalized LU factorization that is required for various algorithms such as PCA (see issue #8049).