@pearu 2019-10-29
@ezyang , the PR is complete as it is, however, there is a magma bug issue, see
  https://github.com/pytorch/pytorch/pull/28608#discussion_r339839370
that appears only for small square arrays on cuda devices. @vishwakftw had concerns that users might interpret the appearing nans as their errors and, as I understood, suggested to hold back this PR. I believe that the issue should be fixed in magma and until then document the magma bug in lu documentation. If that would be satisfactory, I can update the PR and remove WIP asap.
.