@vishwakftw 2019-10-29
@pearu, I think the PR can land as is, but raise an issue when the matrices are singular i.e., preserving older behavior. Perhaps when the magma bug is fixed, we can enable accepting singular matrices. Do you think this is reasonable?