@pearu 2019-10-29
@vishwakftw, let me sum up the situation:
The old behavior is:
1. LU factorization accepts only square matrices
2. LU factorization raises an error when the input matrix is singular

This PR changes the old behavior as follows:
A. LU factorization accepts square and non-square matrices
B. LU factorization allows singular square matrices
C. LU factorization allows singular non-square matrices

The MAGMA bug is effective only when the following conditions are satisfied *all at the same time*:
(i) `A` data is stored on a CUDA device
(ii) `len(A.shape) > 2`, that is, the input `A` contains batches of matrices
(iii) `m == n` and `m <= 32`, that is, the batches contain small square matrices
(iv) `pivot == True`

To preserve the backward compatibility behavior, avoid the MAGMA bug, and support non-square LU factorizations in full, it would be sufficient to enable the following features from this PR:
- A, B, C if storage is on cpu RAM
- A, B, C if storage is on cuda RAM and `len(A.shape) == 2`
- A, C if storage is cuda RAM and `len(A.shape) > 2`, B is disabled (that is, singular square matrices cause an exception when used in patches) until a fix to MAGMA bug will be available.

In essence, the current NaN results will be replaced with an exception until the NaN results will be made impossible by a fix to the MAGMA bug.

Notice that I tend to insist that singular inputs should be allowed as much as possible because for linear algebra algorithms such as truncated SVD or PCA, not allowing singular inputs to LU would be a big blocker. In my opinion, the MAGMA bug is effective only for a small subset of possible application and it would not be reasonable to hold back the development of new algorithms for which there is high demand.

Btw, exists also another approach: instead of using `magmaLuBatched`, use the same approach as in the case of CPU RAM storage. That will reduce performance for certain (but possibly important) cases but would enable A, B, C in-full.