@pearu 2019-10-30
> This approach SGTM!

@ezyang Just to be sure, which approach are you referring to?
1. Raise an exception if a singular square matrix is used in patches. (A, C, partial B)
2. Discard using `magmaLuBatched` and use the CPU algorithm. (A, B, C, all in-full)