@vishwakftw 2019-10-30
@pearu maybe there could another way to resolve it, and this won't affect runtime for non-singular matrices.

I propose the following:
- perform batched LU using `magmaLuBatched`.
- Find if there are any singular matrices
- If yes, replace `nan`s with zero, and fix the pivot tensor to be `arange(1, n + 1)`.