@pearu 2019-10-30
I like the proposal, @vishwakftw, as it will be a full solution.
However, I am not sure that replacement `nan -> 0` will always give the correct LU factorization.
It might be safer to recompute the LU factorization using `magmaLu` for singular matrices.

Re `pivot == arange(1, n + 1)`, that might not be generally correct. In the case of recomputing LU, it would also be unnecessary.