@vishwakftw 2019-10-31
@pearu we could also do that: identify singular matrices using `infos` and perform the LU decomposition for them separately. This would solve the pivot issue too.