@pearu 2019-10-31
yes, exactly, I am currently working on this approach.