@pearu 2019-11-01
Theoretically, there should not be as the loop over `infos` tensor in `batchCheckErrors` was replaced with another loop over `infos` tensor for checking the singularity.
However, I noticed that `batchCheckErrors` copies `infos` to host ram before running the loop while the loop for singularity checking does not do that. This means a performance penalty as the `infos` tensor is accessed item-wise from cuda memory...
I'll fix it and run some performance tests as well.