@pearu 2019-11-01
For the cases where magma issue 13 is effective, there exists some regression in the performance that is related to the overhead of scanning the info tensor for positive values. For instance, consider an LU factorization of an identity matrix in batches:
```
import torch
N = <number of batches>
n = <matrix size>
a=torch.zeros((N, n, n)); a[:]=torch.eye(n); a_cuda=a.cuda()
timing = <timeit a_cuda.lu(pivot=True)>
```
We have the following timing results (best of three runs):

N,n   | pytorch master | this PR
--|----------------------|------------
10,2 | 108 µs ± 3.7 µs | 120 µs ± 425 ns
10000, 2 |154 µs ± 200 ns |183 µs ± 4.17 µs
10000, 32|567 µs ± 2.48 µs|592 µs ± 445 ns
10000, 33|2.39 ms ± 1.8 µs|2.39 ms ± 732 ns

So, the additional overhead is about 12-30 µs per lu call iff magma issue 13 is effective.
Otherwise (see the last row), there is no regression in performance.

Btw, the timings were identical when using `a=torch.randn((N, n, n))`