@vishwakftw 2019-11-04
Semi-automatic rebase is not working, could you please try to rebase manually?