@pearu 2019-11-04
@pytorchbot rebase this please